+++ packages/alfa-trampoline/src/trampoline.ts
@@ -85,13 +85,13 @@ export namespace Trampoline {
 
   export function traverse<T, U>(
     values: Iterable<T>,
-    mapper: Mapper<T, Trampoline<U>>
+    mapper: Mapper<T, Trampoline<U>, [index: number]>
   ): Trampoline<Iterable<U>> {
     return Iterable.reduce(
       values,
-      (values, value) =>
+      (values, value, i) =>
         values.flatMap((values) =>
-          mapper(value).map((value) => Array.append(values, value))
+          mapper(value, i).map((value) => Array.append(values, value))
         ),
       done(Array.empty())
     );